pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "artifactminer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "textual>=0.62",
  "pydantic>=2",
  "httpx>=0.27",
  "fastapi",
  "uvicorn",
  "openai>=2.3.0",
  "python-dotenv>=1.1.1",
  "gitpython",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "aiosqlite>=0.19",
  "email_validator>=2.3.0",
  "python-multipart>=0.0.20",
  "ollama>=0.6.1",
  "pypdf",
  "requests",
  "huggingface-hub>=0.24",
  "typer>=0.12",
  "rich>=13",
  "tree-sitter>=0.24",
  "tree-sitter-python>=0.23",
  "tree-sitter-javascript>=0.23",
  "tree-sitter-typescript>=0.23"
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[project.scripts]
artifactminer = "artifactminer.main:main"
artifactminer-tui = "artifactminer.tui.app:run"
api = "artifactminer.api.main:main"
resume = "artifactminer.resume.cli:main"

# Tell setuptools we're using the src/ layout
[tool.setuptools.packages.find]
where = ["src"]

# Pytest config
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
asyncio_mode = "auto"

# Optional dev extras
[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.21",
  "ruff>=0.6",
  "mypy>=1.10",
]
